Support code update

+++ trunk/SMS/NeoMythMenu/util.h
@@ -18,6 +18,6 @@ extern void strcat_asm(char* dst,const char* src);
 extern void strncat_asm(char* dst,const char* src,BYTE cnt);
 extern const char* get_file_extension_asm(const char* src);
 extern void memcpy_asm(char* dst,const char* src,WORD size);
-extern void memset_asm(char* dst,char val,BYTE size);
+extern void memset_asm(char* dst,char val,WORD size);
 #endif
